refactor(memories): move the page folder to views #181
