+++ src/views/pages/memories/memory/page.connected.js
